% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_str.R
\name{matrix_to_string}
\alias{matrix_to_string}
\title{Convert a Matrix to a Tree String}
\usage{
matrix_to_string(matrix)
}
\arguments{
\item{matrix}{A matrix containing columns:

1. ancestor

2. son1

3. son2

4. brlen

5. name}
}
\value{
A string containing the rooted tree string for the input matrix.
}
\description{
This function allows users to convert a matrix to a tree string in Newick format.
}
\examples{
matrix <- transRate:: sample_matrix
matrix
matrix_to_string(matrix)

}
\seealso{
transRate:: sample_matrix
}
