% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_sim.R
\name{migration_sim}
\alias{migration_sim}
\title{Simulate Migration Events in the Coalescent}
\usage{
migration_sim(migration_rate, treestr, popnames, popnum)
}
\arguments{
\item{migration_rate}{A numerical value containing the desired migration rate.}

\item{treestr}{A tree string, written in Newick format, containing the
coalescent to undergo the migration simulation.}

\item{popnames}{The names of populations in the tree as a list.}

\item{popnum}{The number of populations in the tree as an interger.}
}
\value{
Matrix containing the transformed populations and ancestral nodes.
}
\description{
This function allows users to simulate migration events in a coalescent. The migration rate is drawn from a binomial distribution using the migration_rate parameter.
}
\examples{
treestr <- "(((H1:0.00402#0.01,C1:0.00402#0.01):0.00304#0.01,G2:0.00707#0.01):0.00929#0.01,C2:0.01635#0.01)#0.01;"
migration_rate <- 0.5
popnames <- c("H","C","G")
popnum <- 3
migration_sim(migration_rate,treestr,popnames,popnum)

}
