% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_treestr.R
\docType{data}
\name{sample_treestr}
\alias{sample_treestr}
\title{Sample Data Set with 3 Clades}
\format{
A string with 4919 characters:
}
\source{
{Created by Skylar Ann Gay to serve as an example.}
}
\usage{
sample_treestr
}
\description{
This sample tree string is in Newick fromat and contains 2 populations named "G" and "H" with 100 individuals in each population.
}
\examples{
treestr <- sample_treestr
treestr
}
\keyword{datasets}
